day1 - completed
It's a matter of reading the file and discretising between the lines that are empty and the ones that contain an int. Add them together, chuck them in a list and use the max() function to obtain the desired result.
The second half of the puzzle requires was solved with a loop that find the current max, adds it to a counter and removes the element from the list. At the next iteration, the new max will be different.